volumes ans snapshots lists should use the new pager
